backend/requirements.txt
Django>=4.2.17,<4.3
djangorestframework>=3.15.2,<3.16
djangorestframework-simplejwt>=5.3.1,<5.4
psycopg2>=2.9.10,<2.11
django-cors-headers>=4.6.0,<4.7
django-environ>=0.11.2,<0.12
razorpay>=1.4.2,<1.5
